+++ history.md
@@ -3,6 +3,7 @@
 
 #### v0.11.1
 - Small fix in class `meico.mei.Mei2MusicXmlConverter` to address issue #30.
+- Added method `meico.midi.Midi.append()` to concatenate MIDI sequences.
 
 
 #### v0.11.0